requirements.txt

streamlit>=1.44.0
openai>=1.0.0
python-dotenv
requests
